Add files via upload
Here is one of academic projects that I've participated in. The code is self-explanatory, feel free to study it!

The main focus of the project is researching real estate market in Tehran using the dataset from Kaggle! Our team conducted an analysis of the input data and attempted to build regression models that best describe our data. Additionally, as part of the project, we proposed and tested several hypotheses. The project is educational in nature!